[Unchained] Restructure SAIL (#362)
Some cleanup to clarify the state of the repository at this time:
* Rename the `magic` package `enrichment` instead, since it now only
relates to enriching SPARQL queries, rather than any custom indexing
* Rename `CpNotifyingSail` to `EnrichingSail` since that describes the
purpose better
* Some more associated renaming and moving of things
