+++ src/test/scala/test/services/sparql/regression/TestDb.scala
@@ -17,7 +17,8 @@ import se.lu.nateko.cp.meta.ingestion.{BnodeStabilizers, Ingestion, RdfXmlFileIn
 import se.lu.nateko.cp.meta.instanceserver.Rdf4jInstanceServer
 import se.lu.nateko.cp.meta.services.Rdf4jSparqlRunner
 import se.lu.nateko.cp.meta.services.citation.{CitationClient, CitationProvider, CitationStyle}
-import se.lu.nateko.cp.meta.services.sparql.magic.{CpNotifyingSail, StorageSail}
+import se.lu.nateko.cp.meta.services.sparql.StorageSail
+import se.lu.nateko.cp.meta.services.sparql.enrichment.EnrichingSail
 import se.lu.nateko.cp.meta.utils.async.executeSequentially
 import se.lu.nateko.cp.meta.{LmdbConfig, RdfStorageConfig}
 
@@ -130,7 +131,7 @@ private def makeSail(dir: Path)(using ExecutionContext)(using system: ActorSyste
 
 	val citer = new CitationProvider(base, _ => CitationClientDummy, metaConf)
 	import TestRepo.given
-	CpNotifyingSail(base, citer)
+	EnrichingSail(base, citer)
 }
 
 object CitationClientDummy extends CitationClient {
